+++ source/slang/ir-specialize.h
@@ -1,0 +1,13 @@
+// ir-specialize.h
+#pragma once
+
+namespace Slang
+{
+struct IRModule;
+
+// Find suitable uses of the `specialize` instruction that
+// can be replaced with references to specialized functions.
+void specializeGenerics(
+    IRModule*   module);
+
+}
