+++ source/slang/slang.vcxproj
@@ -181,18 +181,22 @@
     <ClInclude Include="expr-defs.h" />
     <ClInclude Include="glsl.meta.slang.h" />
     <ClInclude Include="hlsl.meta.slang.h" />
+    <ClInclude Include="ir-clone.h" />
     <ClInclude Include="ir-constexpr.h" />
     <ClInclude Include="ir-dce.h" />
     <ClInclude Include="ir-dominators.h" />
     <ClInclude Include="ir-existential.h" />
+    <ClInclude Include="ir-glsl-legalize.h" />
     <ClInclude Include="ir-inst-defs.h" />
     <ClInclude Include="ir-insts.h" />
+    <ClInclude Include="ir-link.h" />
     <ClInclude Include="ir-missing-return.h" />
     <ClInclude Include="ir-restructure-scoping.h" />
     <ClInclude Include="ir-restructure.h" />
     <ClInclude Include="ir-sccp.h" />
     <ClInclude Include="ir-serialize.h" />
     <ClInclude Include="ir-specialize-resources.h" />
+    <ClInclude Include="ir-specialize.h" />
     <ClInclude Include="ir-ssa.h" />
     <ClInclude Include="ir-validate.h" />
     <ClInclude Include="ir.h" />
@@ -232,17 +236,21 @@
     <ClCompile Include="diagnostics.cpp" />
     <ClCompile Include="dxc-support.cpp" />
     <ClCompile Include="emit.cpp" />
+    <ClCompile Include="ir-clone.cpp" />
     <ClCompile Include="ir-constexpr.cpp" />
     <ClCompile Include="ir-dce.cpp" />
     <ClCompile Include="ir-dominators.cpp" />
     <ClCompile Include="ir-existential.cpp" />
+    <ClCompile Include="ir-glsl-legalize.cpp" />
     <ClCompile Include="ir-legalize-types.cpp" />
+    <ClCompile Include="ir-link.cpp" />
     <ClCompile Include="ir-missing-return.cpp" />
     <ClCompile Include="ir-restructure-scoping.cpp" />
     <ClCompile Include="ir-restructure.cpp" />
     <ClCompile Include="ir-sccp.cpp" />
     <ClCompile Include="ir-serialize.cpp" />
     <ClCompile Include="ir-specialize-resources.cpp" />
+    <ClCompile Include="ir-specialize.cpp" />
     <ClCompile Include="ir-ssa.cpp" />
     <ClCompile Include="ir-validate.cpp" />
     <ClCompile Include="ir.cpp" />
